feat: refactor preprocessing into a new service #11731
